refactor: 重構 VideoTrimmer 元件，將功能拆分為子元件以提升可讀性和可維護性，並優化行動裝置的縮圖生成邏輯。
