+++ src/components/pages/VideoUpload/index.tsx
@@ -1,44 +1,20 @@
-import type React from 'react';
-
 import { useState, useRef, useEffect } from 'react';
-import { Button } from '@/components/ui/button';
-import { Slider } from '@/components/ui/slider';
-import {
-  Play,
-  Pause,
-  Check,
-  Plus,
-  Trash2,
-  ChevronLeft,
-  ChevronRight,
-  Image as ImageIcon,
-  AlertCircle,
-} from 'lucide-react';
-import { cn } from '@/lib/utils';
+import { AlertCircle } from 'lucide-react';
 import StepIndicator from '@/components/common/StepIndicator';
 import { useRouter } from 'next/router';
 import { uploadRecipeVideo, updateRecipeSteps } from '@/services/recipes';
+
+// 引入子元件
 import UploadArea from './UploadArea';
+import VideoPlayer from './VideoPlayer';
+import TrimControls from './TrimControls';
+import SegmentNavigation from './SegmentNavigation';
+import SegmentDescription from './SegmentDescription';
+import ActionButtons from './ActionButtons';
 
-// 定義片段類型
-type Segment = {
-  id: string;
-  startTime: number;
-  endTime: number;
-  startPercent: number;
-  endPercent: number;
-  description: string;
-};
-
-// 影片剪輯器元件屬性
-type VideoTrimmerProps = {
-  onSave: (trimmedVideo: {
-    file: File;
-    segments: Segment[];
-    description?: string;
-  }) => void;
-  onCancel: () => void;
-};
+// 引入類型和工具函數
+import { Segment, VideoTrimmerProps, ErrorState } from './types';
+import { generateId, isMobileDevice } from './utils';
 
 /**
  * 影片剪輯器元件，用於上傳、預覽和剪輯影片
@@ -67,10 +43,7 @@ export default function VideoTrimmer({ onSave, onCancel }: VideoTrimmerProps) {
   const [thumbnails, setThumbnails] = useState<string[]>([]);
 
   // 驗證狀態
-  const [errors, setErrors] = useState<{
-    video?: string;
-    description?: string;
-  }>({});
+  const [errors, setErrors] = useState<ErrorState>({});
   const [isSubmitting, setIsSubmitting] = useState<boolean>(false);
 
   // API 錯誤狀態
@@ -294,53 +267,7 @@ export default function VideoTrimmer({ onSave, onCancel }: VideoTrimmerProps) {
   };
 
   /**
-   * 影片載入後的處理，在影片載入後生成縮圖
-   */
-  const atVideoLoaded = () => {
-    if (videoRef.current && videoUrl) {
-      console.log('影片載入完成，初始化片段');
-      const videoDuration = videoRef.current.duration || 134.63;
-      setDuration(videoDuration);
-
-      // 強制再次清除錯誤
-      setErrors({});
-
-      // 更新片段的時長
-      const updatedSegments =
-        segments.length > 0
-          ? segments.map((segment, index) =>
-              index === 0
-                ? { ...segment, endTime: videoDuration, endPercent: 100 }
-                : segment,
-            )
-          : [
-              {
-                id: generateId(),
-                startTime: 0,
-                endTime: videoDuration,
-                startPercent: 0,
-                endPercent: 100,
-                description: '',
-              },
-            ];
-
-      setSegments(updatedSegments);
-      setCurrentSegmentIndex(0);
-      setTrimValues([0, 100]);
-
-      // 針對移動裝置最佳化：只在非行動裝置上生成縮圖，或減少縮圖數量
-      const isMobile = /iPhone|iPad|iPod|Android/i.test(navigator.userAgent);
-      if (!isMobile) {
-        generateThumbnails();
-      } else {
-        // 行動裝置上只生成少量縮圖或跳過
-        generateSimplifiedThumbnails();
-      }
-    }
-  };
-
-  /**
-   * 新增：簡化版縮圖生成函數
+   * 簡化版縮圖生成函數 (行動裝置優化)
    */
   const generateSimplifiedThumbnails = async () => {
     if (!videoRef.current || !videoUrl) return;
@@ -410,10 +337,48 @@ export default function VideoTrimmer({ onSave, onCancel }: VideoTrimmerProps) {
   };
 
   /**
-   * 生成唯一ID，用於標識片段
+   * 影片載入後的處理，在影片載入後生成縮圖
    */
-  const generateId = () => {
-    return Math.random().toString(36).substring(2, 9);
+  const atVideoLoaded = () => {
+    if (videoRef.current && videoUrl) {
+      console.log('影片載入完成，初始化片段');
+      const videoDuration = videoRef.current.duration || 134.63;
+      setDuration(videoDuration);
+
+      // 強制再次清除錯誤
+      setErrors({});
+
+      // 更新片段的時長
+      const updatedSegments =
+        segments.length > 0
+          ? segments.map((segment, index) =>
+              index === 0
+                ? { ...segment, endTime: videoDuration, endPercent: 100 }
+                : segment,
+            )
+          : [
+              {
+                id: generateId(),
+                startTime: 0,
+                endTime: videoDuration,
+                startPercent: 0,
+                endPercent: 100,
+                description: '',
+              },
+            ];
+
+      setSegments(updatedSegments);
+      setCurrentSegmentIndex(0);
+      setTrimValues([0, 100]);
+
+      // 針對移動裝置最佳化：只在非行動裝置上生成縮圖，或減少縮圖數量
+      if (!isMobileDevice()) {
+        generateThumbnails();
+      } else {
+        // 行動裝置上只生成少量縮圖或跳過
+        generateSimplifiedThumbnails();
+      }
+    }
   };
 
   /**
@@ -697,10 +662,7 @@ export default function VideoTrimmer({ onSave, onCancel }: VideoTrimmerProps) {
       hasDescription: segments[currentSegmentIndex]?.description?.length || 0,
     });
 
-    const newErrors: {
-      video?: string;
-      description?: string;
-    } = {};
+    const newErrors: ErrorState = {};
 
     // 驗證影片
     if (!videoFile) {
@@ -831,13 +793,6 @@ export default function VideoTrimmer({ onSave, onCancel }: VideoTrimmerProps) {
   };
 
   /**
-   * 格式化時間顯示 (秒)
-   */
-  const formatTime = (timeInSeconds: number) => {
-    return timeInSeconds.toFixed(2);
-  };
-
-  /**
    * 取消上傳並返回
    */
   const atCancel = () => {
@@ -869,13 +824,41 @@ export default function VideoTrimmer({ onSave, onCancel }: VideoTrimmerProps) {
     }
   }, [segments, currentSegmentIndex]);
 
-  // 獲取當前片段
-  const currentSegment = segments[currentSegmentIndex] || {
-    startTime: 0,
-    endTime: 0,
-    startPercent: 0,
-    endPercent: 0,
-    description: '',
+  // 顯示開發環境調試按鈕
+  const renderDebugButton = () => {
+    if (process.env.NODE_ENV === 'development') {
+      return (
+        <button
+          onClick={() => {
+            console.log('Debug 資訊:', {
+              videoFile: !!videoFile,
+              videoUrl: !!videoUrl,
+              segments: segments.length > 0 ? segments.length : '無片段',
+              currentSegmentIndex,
+              errors: Object.keys(errors).length > 0 ? errors : '無錯誤',
+              isSubmitting,
+              recipeId,
+              description:
+                segments[currentSegmentIndex]?.description || '無說明文字',
+              descriptionLength:
+                segments[currentSegmentIndex]?.description?.trim().length || 0,
+            });
+
+            // 強制清除所有錯誤
+            setErrors({});
+
+            // 延遲重新驗證以確保狀態更新完成
+            setTimeout(() => {
+              validateForm();
+            }, 100);
+          }}
+          className="w-full bg-yellow-50 text-yellow-700 border-yellow-300 p-2 rounded mt-2"
+        >
+          Debug 工具 (重置錯誤狀態)
+        </button>
+      );
+    }
+    return null;
   };
 
   return (
@@ -896,321 +879,52 @@ export default function VideoTrimmer({ onSave, onCancel }: VideoTrimmerProps) {
         />
       ) : (
         <div className="px-4 space-y-4">
-          {/* 影片預覽 */}
-          <div className="relative aspect-video bg-gray-100 rounded-lg overflow-hidden">
-            {videoUrl ? (
-              <video
-                ref={videoRef}
-                src={videoUrl}
-                className="w-full h-full object-contain"
-                onLoadedMetadata={atVideoLoaded}
-                onTimeUpdate={atTimeUpdate}
-                onClick={atTogglePlayPause}
-                playsInline
-                preload="metadata"
-                crossOrigin="anonymous"
-              >
-                <track kind="captions" label="中文" default />
-                您的瀏覽器不支援影片標籤
-              </video>
-            ) : (
-              <div className="flex items-center justify-center h-full">
-                <ImageIcon
-                  className="h-12 w-12 text-gray-400"
-                  aria-hidden="true"
-                />
-              </div>
-            )}
-
-            {/* 播放/暫停按鈕覆蓋層 */}
-            {videoUrl && (
-              <div className="absolute inset-0 flex items-center justify-center">
-                <Button
-                  variant="ghost"
-                  size="icon"
-                  className="h-16 w-16 rounded-full bg-black/30 text-white hover:bg-black/50"
-                  onClick={atTogglePlayPause}
-                >
-                  {isPlaying ? (
-                    <Pause className="h-8 w-8" />
-                  ) : (
-                    <Play className="h-8 w-8" />
-                  )}
-                </Button>
-              </div>
-            )}
-          </div>
-
-          {/* 時間顯示 */}
-          <div className="flex justify-between px-4 py-2 text-sm">
-            <div>當前: {formatTime(currentTime)} 秒</div>
-            <div>總長: {formatTime(duration)} 秒</div>
-          </div>
+          {/* 影片播放器 */}
+          <VideoPlayer
+            videoUrl={videoUrl}
+            isPlaying={isPlaying}
+            currentTime={currentTime}
+            duration={duration}
+            videoRef={videoRef}
+            atTogglePlayPause={atTogglePlayPause}
+            atTimeUpdate={atTimeUpdate}
+            atVideoLoaded={atVideoLoaded}
+          />
 
           {/* 片段導航 */}
-          <div className="flex items-center justify-between px-4 py-2">
-            <button
-              onClick={atGoPreviousSegment}
-              className="p-2 text-gray-600"
-              disabled={currentSegmentIndex === 0}
-            >
-              <ChevronLeft className="h-5 w-5" />
-            </button>
-            <div className="text-sm">
-              步驟 {currentSegmentIndex + 1}/{segments.length}
-            </div>
-            <button
-              onClick={atGoNextSegment}
-              className="p-2 text-gray-600"
-              disabled={currentSegmentIndex === segments.length - 1}
-            >
-              <ChevronRight className="h-5 w-5" />
-            </button>
-            <button onClick={atTogglePlayPause} className="p-2 text-gray-600">
-              {isPlaying ? (
-                <Pause className="h-5 w-5" />
-              ) : (
-                <Play className="h-5 w-5" />
-              )}
-            </button>
-            <button onClick={atAddSegment} className="p-2 text-gray-600">
-              <Plus className="h-5 w-5" />
-            </button>
-          </div>
+          <SegmentNavigation
+            segments={segments}
+            currentSegmentIndex={currentSegmentIndex}
+            isPlaying={isPlaying}
+            atGoPreviousSegment={atGoPreviousSegment}
+            atGoNextSegment={atGoNextSegment}
+            atTogglePlayPause={atTogglePlayPause}
+            atAddSegment={atAddSegment}
+            atDeleteCurrentSegment={atDeleteCurrentSegment}
+          />
 
-          {/* 雙滑桿剪輯 (YouTube 風格) */}
-          <div className="space-y-2 mt-6">
-            {/* 時間標記 */}
-            <div className="flex justify-between text-xs text-gray-500">
-              <span>0:00</span>
-              <span>
-                {Math.floor(duration / 60)}:
-                {Math.floor(duration % 60)
-                  .toString()
-                  .padStart(2, '0')}
-              </span>
-            </div>
-
-            {/* 整合縮圖預覽和滑桿 */}
-            <div className="relative h-16">
-              {/* 縮圖容器 */}
-              <div className="absolute inset-0 flex rounded overflow-hidden">
-                {thumbnails.map((thumbnail, i) => {
-                  // 生成唯一且穩定的 key，使用縮圖資料的前10字元作為唯一標識
-                  const uniqueId = `${i}-${thumbnail.slice(-10)}`;
-                  return (
-                    <div
-                      key={uniqueId}
-                      className="h-full flex-grow"
-                      style={{ width: `${100 / thumbnails.length}%` }}
-                    >
-                      <img
-                        src={thumbnail || '/placeholder.svg'}
-                        alt={`影片縮圖 ${i + 1}`}
-                        className="h-full w-full object-cover"
-                      />
-                    </div>
-                  );
-                })}
-              </div>
-
-              {/* 所有片段標記 */}
-              {segments.map((segment, index) => (
-                <div
-                  key={segment.id}
-                  className={cn(
-                    'absolute top-0 bottom-0 border-2 pointer-events-none z-10',
-                    index === currentSegmentIndex
-                      ? 'border-blue-500 bg-blue-500/10'
-                      : 'border-blue-300 bg-blue-300/10',
-                  )}
-                  style={{
-                    left: `${segment.startPercent}%`,
-                    width: `${segment.endPercent - segment.startPercent}%`,
-                  }}
-                >
-                  <div className="absolute -top-5 left-1/2 -translate-x-1/2 text-xs font-medium px-1 py-0.5 rounded bg-blue-600 text-white">
-                    {index + 1}
-                  </div>
-                </div>
-              ))}
-
-              {/* 非選中區域遮罩 (左側) */}
-              <div
-                className="absolute top-0 bottom-0 bg-black/50 pointer-events-none z-10 rounded-l"
-                style={{
-                  left: 0,
-                  width: `${trimValues[0]}%`,
-                }}
-              />
-
-              {/* 非選中區域遮罩 (右側) */}
-              <div
-                className="absolute top-0 bottom-0 bg-black/50 pointer-events-none z-10 rounded-r"
-                style={{
-                  right: 0,
-                  width: `${100 - trimValues[1]}%`,
-                }}
-              />
-
-              {/* 當前播放位置指示器 */}
-              <div
-                className="absolute top-0 bottom-0 w-0.5 bg-red-500 z-20"
-                style={{
-                  left: `${(currentTime / duration) * 100}%`,
-                }}
-              >
-                <div className="absolute -top-1 -ml-1.5 w-3 h-3 bg-red-500 rounded-full" />
-              </div>
-
-              {/* 自定義樣式的 Slider */}
-              <Slider
-                value={trimValues as [number, number]}
-                min={0}
-                max={100}
-                step={0.1}
-                onValueChange={atTrimChange}
-                className="absolute inset-0 z-30 [&_[data-orientation=horizontal]]:bg-transparent [&_[role=slider]]:opacity-0 [&_[data-orientation=horizontal]>.range]:bg-transparent"
-                thumbClassName="group h-full w-1 top-0 rounded-none -mt-0 bg-transparent"
-              />
-
-              {/* 左側把手 (與 Slider 的第一個滑塊對齊) */}
-              <div
-                className="absolute top-0 bottom-0 w-1 bg-blue-500 z-40 pointer-events-none"
-                style={{ left: `${trimValues[0]}%` }}
-              >
-                <div className="absolute h-6 w-6 bg-blue-500 rounded-full -ml-3 top-1/2 -translate-y-1/2 flex items-center justify-center">
-                  <div className="h-4 w-0.5 bg-white" />
-                </div>
-              </div>
-
-              {/* 右側把手 (與 Slider 的第二個滑塊對齊) */}
-              <div
-                className="absolute top-0 bottom-0 w-1 bg-blue-500 z-40 pointer-events-none"
-                style={{ left: `${trimValues[1]}%` }}
-              >
-                <div className="absolute h-6 w-6 bg-blue-500 rounded-full -ml-3 top-1/2 -translate-y-1/2 flex items-center justify-center">
-                  <div className="h-4 w-0.5 bg-white" />
-                </div>
-              </div>
-            </div>
-          </div>
-
-          {/* 剪輯時間顯示 */}
-          <div className="flex justify-between text-sm">
-            <span>起點: {formatTime(currentSegment.startTime)} 秒</span>
-            <span>終點: {formatTime(currentSegment.endTime)} 秒</span>
-          </div>
-
-          {/* 標記按鈕 */}
-          <div className="flex justify-between gap-4 mt-4">
-            <Button
-              onClick={atMarkStartPoint}
-              variant="outline"
-              className="flex-1"
-            >
-              標記起點
-            </Button>
-            <Button
-              onClick={atMarkEndPoint}
-              variant="outline"
-              className="flex-1"
-            >
-              標記終點
-            </Button>
-          </div>
-
-          {/* 紅綠燈警示 */}
-          <div className="mt-3 mb-1">
-            {(() => {
-              const segmentDuration =
-                currentSegment.endTime - currentSegment.startTime;
-              let statusColor = '';
-              let statusText = '';
-
-              if (segmentDuration < 5 || segmentDuration > 30) {
-                statusColor = 'bg-red-500';
-                statusText =
-                  segmentDuration < 5
-                    ? '時間太短 (建議至少5秒)'
-                    : '時間太長 (建議不超過30秒)';
-              } else if (
-                (segmentDuration >= 5 && segmentDuration < 10) ||
-                (segmentDuration > 25 && segmentDuration <= 30)
-              ) {
-                statusColor = 'bg-yellow-500';
-                statusText =
-                  segmentDuration < 10
-                    ? '時間略短 (適中為10-25秒)'
-                    : '時間略長 (適中為10-25秒)';
-              } else {
-                statusColor = 'bg-green-500';
-                statusText = '時間長度適中';
-              }
-
-              return (
-                <div className="flex items-center">
-                  <div className={`w-4 h-4 rounded-full ${statusColor} mr-2`} />
-                  <div className="text-sm text-gray-700">
-                    <span>步驟時長: {segmentDuration.toFixed(2)} 秒</span>
-                    <span className="ml-2 text-xs text-gray-500">
-                      {statusText}
-                    </span>
-                  </div>
-                </div>
-              );
-            })()}
-          </div>
-
-          {/* 重置按鈕 */}
-          <Button
-            onClick={atResetCurrentSegment}
-            variant="outline"
-            className="w-full flex items-center justify-center"
-          >
-            <span className="mr-2">↻</span>
-            該步驟重置
-          </Button>
-
-          {/* 刪除步驟按鈕 */}
-          <Button
-            onClick={atDeleteCurrentSegment}
-            variant="outline"
-            disabled={segments.length <= 1}
-            className="w-full bg-gray-200 text-gray-700 rounded-md py-2 flex items-center justify-center mt-2"
-          >
-            <Trash2 className="h-4 w-4 mr-2" />
-            刪除此步驟
-          </Button>
+          {/* 剪輯控制 */}
+          <TrimControls
+            duration={duration}
+            currentTime={currentTime}
+            trimValues={trimValues}
+            thumbnails={thumbnails}
+            atTrimChange={atTrimChange}
+            atMarkStartPoint={atMarkStartPoint}
+            atMarkEndPoint={atMarkEndPoint}
+            atResetCurrentSegment={atResetCurrentSegment}
+            segments={segments}
+            currentSegmentIndex={currentSegmentIndex}
+          />
 
           {/* 說明文字 */}
-          <div className="mt-4">
-            <h3 className="text-sm font-medium text-gray-700 mb-2">
-              說明文字 (步驟 {currentSegmentIndex + 1})
-              <span className="ml-2 text-xs text-gray-500">
-                {segments[currentSegmentIndex]?.description?.trim().length || 0}
-                /10 字元
-              </span>
-            </h3>
-            <textarea
-              value={segments[currentSegmentIndex]?.description || ''}
-              onChange={atDescriptionChange}
-              onBlur={() => validateForm()}
-              className={cn(
-                'w-full p-2 text-sm border rounded-md min-h-[80px] resize-none',
-                errors.description
-                  ? 'border-red-500 bg-red-50'
-                  : 'border-gray-300 text-gray-600',
-              )}
-              placeholder="請輸入此步驟的說明文字，至少需要10個字元..."
-            />
-            {errors.description && (
-              <div className="text-red-500 text-sm mt-1 flex items-center error-message">
-                <AlertCircle className="h-4 w-4 mr-1" aria-hidden="true" />
-                {errors.description}
-              </div>
-            )}
-          </div>
+          <SegmentDescription
+            segments={segments}
+            currentSegmentIndex={currentSegmentIndex}
+            error={errors.description}
+            atDescriptionChange={atDescriptionChange}
+            validateForm={validateForm}
+          />
 
           {/* API 錯誤訊息 */}
           {apiError && (
@@ -1221,140 +935,19 @@ export default function VideoTrimmer({ onSave, onCancel }: VideoTrimmerProps) {
           )}
 
           {/* Debug 按鈕 - 開發環境使用 */}
-          {process.env.NODE_ENV === 'development' && (
-            <Button
-              onClick={() => {
-                console.log('Debug 資訊:', {
-                  videoFile: !!videoFile,
-                  videoUrl: !!videoUrl,
-                  segments: segments.length > 0 ? segments.length : '無片段',
-                  currentSegmentIndex,
-                  errors: Object.keys(errors).length > 0 ? errors : '無錯誤',
-                  isSubmitting,
-                  recipeId,
-                  description:
-                    segments[currentSegmentIndex]?.description || '無說明文字',
-                  descriptionLength:
-                    segments[currentSegmentIndex]?.description?.trim().length ||
-                    0,
-                });
+          {renderDebugButton()}
 
-                // 強制清除所有錯誤
-                setErrors({});
-
-                // 如果沒有片段但有影片，創建一個預設片段
-                if (segments.length === 0 && videoFile) {
-                  const initialSegment: Segment = {
-                    id: generateId(),
-                    startTime: 0,
-                    endTime: duration,
-                    startPercent: 0,
-                    endPercent: 100,
-                    description:
-                      '這是一段預設的說明文字，至少包含十個字元以上。這應該足以通過驗證了。',
-                  };
-                  setSegments([initialSegment]);
-                  setCurrentSegmentIndex(0);
-                } else if (segments.length > 0) {
-                  // 確保現有片段說明文字足夠長
-                  const updatedSegments = [...segments];
-                  if (
-                    !updatedSegments[currentSegmentIndex]?.description ||
-                    updatedSegments[currentSegmentIndex].description.trim()
-                      .length < 10
-                  ) {
-                    updatedSegments[currentSegmentIndex] = {
-                      ...updatedSegments[currentSegmentIndex],
-                      description:
-                        '這是一段預設的說明文字，至少包含十個字元以上。這應該足以通過驗證了。',
-                    };
-                    setSegments(updatedSegments);
-                  }
-                }
-
-                // 延遲重新驗證以確保狀態更新完成
-                setTimeout(() => {
-                  const valid = validateForm();
-                  console.log('重新驗證結果:', valid);
-                  if (!valid) {
-                    console.log('驗證仍然失敗，再次強制清除錯誤');
-                    setErrors({});
-                  }
-                }, 100);
-              }}
-              variant="outline"
-              className="w-full bg-yellow-50 text-yellow-700 border-yellow-300"
-            >
-              Debug 工具 (重置錯誤狀態)
-            </Button>
-          )}
-
-          {/* 按鈕群組 */}
-          <div className="flex gap-3 mt-6">
-            <Button onClick={atCancel} variant="outline" className="w-1/2">
-              取消
-            </Button>
-            <Button
-              onClick={() => {
-                // 先檢查文字字數，如果已達標準但按鈕仍禁用，強制清除錯誤
-                const currentDescription =
-                  segments[currentSegmentIndex]?.description?.trim() || '';
-                if (
-                  currentDescription.length >= 10 &&
-                  Object.keys(errors).length > 0
-                ) {
-                  console.log('文字已符合標準但按鈕仍禁用，強制清除錯誤');
-                  setErrors({});
-                  setTimeout(() => atSubmit(), 50);
-                } else {
-                  console.log('嘗試提交，當前按鈕狀態:', {
-                    disabled: isSubmitting || Object.keys(errors).length > 0,
-                    isSubmitting,
-                    errorsCount: Object.keys(errors).length,
-                    errors,
-                    descriptionLength: currentDescription.length,
-                  });
-                  atSubmit();
-                }
-              }}
-              disabled={isSubmitting || Object.keys(errors).length > 0}
-              variant={
-                isSubmitting || Object.keys(errors).length > 0
-                  ? 'secondary'
-                  : 'default'
-              }
-              className="w-1/2"
-            >
-              {isSubmitting ? (
-                '上傳中...'
-              ) : (
-                <>
-                  <Check className="h-5 w-5 mr-2" />
-                  完成
-                </>
-              )}
-            </Button>
-          </div>
-
-          {/* 強制重設狀態按鈕 - 當文字已足夠但按鈕仍然禁用時使用 */}
-          {Object.keys(errors).length > 0 &&
-            segments[currentSegmentIndex]?.description?.trim().length >= 10 && (
-              <Button
-                onClick={() => {
-                  console.log('強制重設狀態並提交');
-                  setErrors({});
-                  setTimeout(() => {
-                    if (Object.keys(errors).length === 0) {
-                      atSubmit();
-                    }
-                  }, 100);
-                }}
-                variant="default"
-                className="w-full mt-2 bg-green-600 hover:bg-green-700 text-white"
-              >
-                強制繼續 (狀態已修復)
-              </Button>
-            )}
+          {/* 操作按鈕 */}
+          <ActionButtons
+            segments={segments}
+            currentSegmentIndex={currentSegmentIndex}
+            isSubmitting={isSubmitting}
+            errors={errors}
+            apiError={apiError}
+            atCancel={atCancel}
+            atSubmit={atSubmit}
+            setErrors={setErrors}
+          />
         </div>
       )}
     </div>
